test: consolidate Playwright scraper mocks into test_scrapers/conftest (slim R5)
Hoist the copy-pasted _make_scraper (5x) and _make_mock_page (4x) helpers
into tests/test_scrapers/conftest.py as make_scraper/make_mock_page. Each
file keeps its prior call-site semantics via thin local aliases/wrappers
(is_cdp default, page-only return, query_selector=None for the auth flow).

The fake_goto/fake_fetch closures are intentionally left in place: their
bodies diverge per test (different redirect/branch logic over local state),
so merging would need a multi-arg config helper — against house style.

Behavior-preserving: tests/test_scrapers green (104), full suite green
(1344 passed, 154 skipped), ruff clean. Net -158 LOC across the 5 files.
